begin port of notebook into codebase
